Fix Error

+++ app/src/main/java/com/example/sport_planet/data/model/BoardModel.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 data class BoardModel(
     val boardId: Long,
-    val hostId: Int,
+    val hostId: Long,
     val hostName: String,
     val title: String,
     val groupStatus: GroupStatusModel,
