+++ app/src/main/java/com/example/sport_planet/data/model/HostModel.kt
@@ -2,7 +2,7 @@ package com.example.sport_planet.data.model
 
 data class HostModel(
     val dislikes: Int,
-    val hostId: Int,
+    val hostId: Long,
     val hostName: String,
     val likes: Int
 )
